# UserBalanceService/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A service that can perform some operations with user balances.",
        "title": "User Balance Service",
        "contact": {
            "name": "Evgeny Belonogov",
            "url": "https://www.t.me/evgenchikkkkkk",
            "email": "ewbelonogov@ya.ru"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/report": {
            "post": {
                "description": "Responds with the link to report download page if OK.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Create report according to the selected period",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "ReportSchemaJSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ReportSchemaJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.ReportResponseJSON"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponseJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponseJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponseJSON"
                        }
                    }
                }
            }
        },
        "/report/{filename}": {
            "get": {
                "description": "download report file if OK.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Download report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "download file",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponseJSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponseJSON": {
            "type": "object",
            "properties": {
                "http error": {
                    "type": "string"
                }
            }
        },
        "api.ReportResponseJSON": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                }
            }
        },
        "api.ReportSchemaJSON": {
            "type": "object",
            "required": [
                "period"
            ],
            "properties": {
                "period": {
                    "type": "string",
                    "example": "2022-10"
                }
            }
        },
...
    }
}